docs: update checkbox.mdx
