+++ docs/app/docs/components/[slug]/components/checkbox/basic.demo.tsx
@@ -1,0 +1,40 @@
+'use client'
+
+import { Box } from '@/styled-system/jsx'
+import { Checkbox, CheckboxGroup, For } from '@cerberus-design/react'
+
+const items = [
+  {
+    id: 'terms',
+    label: 'I agree to the terms and conditions',
+    required: true,
+  },
+  {
+    id: 'legal',
+    label: 'I would like to receive marketing emails',
+  },
+]
+
+export function BasicDemo() {
+  return (
+    <Box w="1/2">
+      <CheckboxGroup name="user_acceptance">
+        <For each={items}>
+          {(item) => (
+            <Checkbox
+              key={item.id}
+              ids={{ control: item.id }}
+              required={item?.required}
+            >
+              {item.label}
+            </Checkbox>
+          )}
+        </For>
+
+        <Checkbox ids={{ control: 'mixed' }} checked="indeterminate">
+          The indeterminate state
+        </Checkbox>
+      </CheckboxGroup>
+    </Box>
+  )
+}
