+++ docs/app/docs/components/[slug]/components/checkbox/static.tsx
@@ -1,91 +1,24 @@
-'use client'
+import { BasicDemo } from './basic.demo'
+import { IndeterminateDemo } from './indeterminate.demo'
+import { SizeDemo } from './size.demo'
+import { CustomDemo } from './custom.demo'
 
-import {
-  Checkbox,
-  CheckboxGroup,
-  For,
-  CheckboxParts,
-} from '@cerberus-design/react'
-import { ArrowDownRight } from '@carbon/icons-react'
-import { Box } from 'styled-system/jsx'
-
-export function OverviewPreview() {
-  const items = [
-    {
-      id: 'terms',
-      label: 'I agree to the terms and conditions',
-
-      required: true,
-    },
-    {
-      id: 'legal',
-      label: 'I would like to receive marketing emails',
-    },
-  ]
-
-  return (
-    <Box w="1/2">
-      <CheckboxGroup name="user_acceptance">
-        <For each={items}>
-          {(item) => (
-            <Checkbox
-              key={item.id}
-              ids={{ control: item.id }}
-              required={Boolean(item.required)}
-            >
-              {item.label}
-            </Checkbox>
-          )}
-        </For>
-
-        <Checkbox ids={{ control: 'mixed' }} checked="indeterminate">
-          The indeterminate state
-        </Checkbox>
-      </CheckboxGroup>
-    </Box>
-  )
-}
-
-export function SizeDemo() {
-  return (
-    <Box w="1/2">
-      <CheckboxGroup name="size">
-        <Checkbox ids={{ control: 'small' }} size="sm">
-          Small
-        </Checkbox>
-        <Checkbox ids={{ control: 'medium' }} size="md">
-          Medium
-        </Checkbox>
-      </CheckboxGroup>
-    </Box>
-  )
-}
-
-export function CustomCheckbox() {
-  return (
-    <Box w="1/2">
-      <CheckboxParts.Root>
-        <CheckboxParts.Label>Custom Checkbox</CheckboxParts.Label>
-        <CheckboxParts.Control
-          css={{
-            bgColor: 'black',
-            border: '4px solid',
-            borderColor: 'danger.border.initial',
-            h: 8,
-            rounded: 'none',
-            transform: 'skewX(-10deg)',
-            w: 8,
-            _checked: {
-              bgColor: 'info.bg.initial',
-            },
-          }}
-        >
-          <CheckboxParts.Indicator>
-            <ArrowDownRight />
-          </CheckboxParts.Indicator>
-        </CheckboxParts.Control>
-        <CheckboxParts.HiddenInput />
-      </CheckboxParts.Root>
-    </Box>
-  )
+export const DEMOS = {
+  basic: {
+    id: 'checkbox.basic',
+    preview: <BasicDemo />,
+  },
+  indeterminate: {
+    id: 'checkbox.indeterminate',
+    preview: <IndeterminateDemo />,
+  },
+  size: {
+    id: 'checkbox.size',
+    preview: <SizeDemo />,
+  },
+  custom: {
+    id: 'checkbox.custom',
+    preview: <CustomDemo />,
+  },
+  meta: `import { Checkbox, CheckboxGroup } from '@cerberus/react'`,
 }
